# Improve iridas_cube parsing speed