While investigating long processor creation time (> 3.5sec), I noticed `iridas_cube` was really slow compared to `spi3d` for 3DLUT loading (on my system around 900ms and 250ms respectively for a 65x65x65 LUT). This brings it to similar performance.

Here is a benchmark using a 65x65x65 .cube 3DLUT, processor creation time (mean of 3 runs `ocioperf --transform`):

|                        | main               |  pr                   |
| ------------- | ------------- | ------------- |
| macOS           | 376 ms          | 216 ms           |
| CentOS 7       | 927 ms          | 347 ms          |
| Windows 10   | 1116 ms         | 715 ms           |