@remia 2022-04-27
I'll try to include Windows numbers shortly.